And now, we do not need StarlarkMapping any longer since we can fulfill the exact same interface with LarkyMapping. Delete StarlarkMapping and move the functionality to LarkyMapping
